Don't throw exceptiones when publishing an event
The RuleScheduler did publish a RuleEvaluationFailed event on a network
outage. The handler for that event would also require the network and
caused another exception which was never caught. That caused the
scheduler to terminate and eventually to hang the hole application.
There wasn't even a log message, as errors on the scheduling threads are
eating per default.