Reset Hotwater program to OWNPROGRAM
I do change the hot water program manually so that I can heat it up when
there is enough excess electricity available. However, once I forgot to
change it back to OWNPROGRAM and it heated up during the night, which
lead to a depleted battery. This commit adds a rule to automatically
change the program back to OWNPROGRAM.