Fix bogus unlocking
Weather events cause constant unlocking of those shutters. Unlocking
means moving the shutter into the last desired state. Now, if in between
a manual usage of the buttons happened, that didn't change the desired
states and the shutters would revert back on the next weather event.

Example: Desired state is closed, from the last DuskStarted event. Ute
now opens the shutters manually in the morning, before the DawnStarted
event fired. The shutters are now open, but the desired state is still
CLOSED. Now the new weather event fires and causes an unlock (because of
no wind) which will now revert back to the desired state, i.e. close the
shutters.

This commit fixes this behaviour, by checking if we are acutally locked
or not. In case the shutter was not locked, nothing will happen. I guess
I will need a follow up commit, to actually consider manual changes as
well.