Turn on Dehumidifier when there is excess energy
I turned them off when the heater was on, cause I controlled the heater
as well, which messed up the energy consumption. I.e. it ended up
consuming more than there was excess energy. Now as the heater is not
controlled, this is no more needed.

Also Turning on happens only if the current excess energy is as well
above the threshold (additionally to the window).